change type to int for option values : KP-7286

+++ src/Core/SharedKernels/DataCollection/DataCollection.Portable/CategoricalOption.cs
@@ -2,8 +2,8 @@ namespace WB.Core.SharedKernels.DataCollection
 {
     public class CategoricalOption
     {
-        public long Value { get; set; }
-        public long? ParentValue { get; set; }
+        public int Value { get; set; }
+        public int? ParentValue { get; set; }
         public string Title { get; set; }
     }
 }
